requirements.txt
# Core Libraries
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2

# Deep Learning
timm==0.9.2
transformers==4.30.2
torch-geometric==2.3.1
dgl==1.0.2
torch-scatter==2.1.1
torch-sparse==0.6.17

# Privacy
opacus==1.4.0
cryptography==40.0.2
pysyft==0.8.0
diffprivlib==0.6.0

# Visualization
plotly==5.15.0
dash==2.10.0
reportlab==3.6.13
kaleido==0.2.1

# Data Processing
h5py==3.8.0
pyarrow==12.0.1
pillow==9.5.0

# System Monitoring
psutil==5.9.5
GPUtil==1.4.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
joblib==1.2.0
tensorboard==2.13.0
ipywidgets==8.0.6

# Documentation
sphinx==6.2.1
sphinx-rtd-theme==1.2.1

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development
black==23.3.0
isort==5.12.0
mypy==1.3.0
flake8==6.0.0